new change

+++ Chatgpt practice Questions/Functions/_6CelciustoFahrenheit.java
@@ -2,13 +2,13 @@
  */
 
 public class _6CelciustoFahrenheit {
-      public static int celciusconvertor(int num) {
-         int result = (num*9/5) +32;
+      public static double celciusconvertor(int num) {
+         double result = (num*9/5) +32;
          return result;
       }
 
     public static void main(String[] args) {
-         int number = 35;
+         double number = 35;
          System.out.println(" The tempperature after converting into fahrenheit:"+celciusconvertor(number));
     }
 }
